docs/requirements.txt
sphinx>=4.0
sphinx-book-theme
myst_nb
tqdm
torchvision
ipywidgets